refactor(network): unify LanguageService into LanguageInterceptor for streamlined language handling
- Merged the logic of LanguageService into LanguageInterceptor to centralize language management.
- Removed the need for a separate LanguageService, with the LanguageInterceptor now directly handling locale settings in network requests.
