+++ app/src/main/java/com/daniellapaiva/movieapp/data/remote/api/MovieApiService.kt
@@ -1,4 +1,4 @@
-package com.daniellapaiva.movieapp.data
+package com.daniellapaiva.movieapp.data.remote.api
 
 import com.daniellapaiva.movieapp.data.model.MovieDetailResponse
 import com.daniellapaiva.movieapp.data.model.MovieListResponse
@@ -12,13 +12,11 @@ interface MovieApiService {
     suspend fun getPopularMovies(
         @Query("api_key") apiKey: String,
         @Query("page") page: Int,
-        @Query("language") language: String
     ): MovieListResponse
 
     @GET("movie/{movie_id}")
     suspend fun getMovieDetails(
         @Path("movie_id") movieId: Int,
         @Query("api_key") apiKey: String,
-        @Query("language") language: String
     ): MovieDetailResponse
 }
