+++ app/src/main/java/com/daniellapaiva/movieapp/di/AppModule.kt
@@ -4,7 +4,7 @@ import com.daniellapaiva.movieapp.data.repository.MovieRepositoryImpl
 import com.daniellapaiva.movieapp.domain.usecase.GetPopularMoviesUseCase
 import com.daniellapaiva.movieapp.domain.MovieRepository
 import com.daniellapaiva.movieapp.domain.usecase.GetMovieDetailsUseCase
-import com.daniellapaiva.movieapp.domain.util.LanguageService
+import com.daniellapaiva.movieapp.data.remote.network.LanguageInterceptor
 import com.daniellapaiva.movieapp.presentation.viewmodel.MovieDetailsViewModel
 import com.daniellapaiva.movieapp.presentation.viewmodel.MovieListViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
@@ -12,9 +12,9 @@ import org.koin.dsl.module
 
 val appModule = module {
 
-    single { LanguageService() }
+    single { LanguageInterceptor() }
 
-    single<MovieRepository> { MovieRepositoryImpl(get(), get()) }
+    single<MovieRepository> { MovieRepositoryImpl(get()) }
 
     single { GetPopularMoviesUseCase(get()) }
     single { GetMovieDetailsUseCase(get()) }
