+++ app/src/main/java/com/daniellapaiva/movieapp/di/NetworkModule.kt
@@ -1,11 +1,14 @@
 package com.daniellapaiva.movieapp.di
 
-import com.daniellapaiva.movieapp.data.MovieApiService
+import com.daniellapaiva.movieapp.data.remote.api.MovieApiService
+import com.daniellapaiva.movieapp.data.remote.network.LanguageInterceptor
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import org.koin.dsl.module
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
+
+private const val BASE_URL = "https://api.themoviedb.org/3/"
 
 val networkModule = module {
     single {
@@ -13,12 +16,15 @@ val networkModule = module {
             level = HttpLoggingInterceptor.Level.BODY
         }
 
+        val languageInterceptor = LanguageInterceptor()
+
         val client = OkHttpClient.Builder()
             .addInterceptor(loggingInterceptor)
+            .addInterceptor(languageInterceptor)
             .build()
 
         Retrofit.Builder()
-            .baseUrl("https://api.themoviedb.org/3/")
+            .baseUrl(BASE_URL)
             .addConverterFactory(GsonConverterFactory.create())
             .client(client)
             .build()
